core: add dedicated file for FromTypeToType
